backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# AI/ML dependencies (CPU versions)
torch>=2.1.0
torchvision>=0.16.0
timm==0.9.8
tensorflow==2.13.0
albumentations==1.3.1
onnx==1.14.0
onnx2pytorch==0.4.1
torchsummary==1.5.1

# Multimedia + data processing
moviepy==1.0.3
librosa==0.10.1
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.25.0
wget==3.2
gradio==3.44.4
ffmpeg-python==0.2.0

# Development + utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests>=2.32.0
tqdm>=4.66.0
scikit-learn==1.3.0
scikit-image>=0.21.0

# Hugging Face Transformers for pre-trained models
transformers>=4.40.0
accelerate==0.24.0
huggingface_hub>=0.18.0

# Real deepfake detection models
efficientnet-pytorch==0.7.1
gdown==4.7.1
